+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-##  (2022-07-19)
+##  (2022-07-22)
 
 [View Release](https://github.com/dev-tvo/undermined-two.git/commits/tag/)
 
 *  initial commit *(Thom van Oort)*
+*  - added curse icons - added lazy load & loading icons - new items data with itemType and curse cost - reworked some styling regarding background & page layout *(Thom van Oort)*